# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Backend gestionnaire de mots de passe
  description: >
    API Zero-Knowledge pour un gestionnaire de mots de passe : le serveur ne voit et ne stocke
    jamais le mot de passe maitre d'un utilisateur, ni la cle qui chiffre son coffre. Voir
    docs/API.md pour les explications detaillees (modele d'authentification, invalidation de
    session, rate limiting). La route WebSocket GET /ws n'est pas representable par OpenAPI 3.0
    (pas de support natif) : voir API.md.
  version: "1.0.0"
servers:
  - url: http://localhost:3000
    description: Local / Docker Compose
security: []
...
paths:
...
  /shares/{id}:
    get:
      tags: [sharing]
      summary: Recupere le blob scelle d'un partage (reserve au destinataire)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ShareId' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SharedEntryView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [sharing]
      summary: Revoque un partage (proprietaire ou destinataire, les deux cotes peuvent)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/ShareId' }
      responses:
        '204': { description: Revoque }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token JWT obtenu via /auth/login ou /auth/refresh, a envoyer dans l'en-tete
        "Authorization: Bearer <token>".
  parameters:
...
    ShareId:
      in: path
      name: id
      required: true
      schema: { type: string }
      description: Identifiant (UUID) d'un partage d'entree
  responses:
...
    Unauthorized:
      description: Non authentifie, session expiree, ou invalidee
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Ressource introuvable ou n'appartenant pas a l'appelant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    RateLimited:
      description: Limite de debit depassee
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    SharedEntryView:
      type: object
      properties:
        owner_email: { type: string }
        sealed_entry: { type: string }
...